+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This is a changelog for **ZachranObed** application.
 
+## [0.4.0]
+### Added
+- **ZOB-91** Add user info item.
+
+### Changed
+- **ZOB-105** Add non-prod firebase projects.
+- **ZOB-92** Update toolbar in Profile screen.
+
+### Fixed
+
+### Removed
+
 ## [0.3.0]
 ### Added
 - **ZOB-83** Rebuild Home screen after return to application.
@@ -9,10 +21,6 @@ This is a changelog for **ZachranObed** application.
 ### Changed
 - **ZOB-80** InfoBanner background color customisation and set to success color.
 - **ZOB-90** Update simple menu item with app version by design.
-
-### Fixed
-
-### Removed
 
 ## [0.2.0]
 ### Added